Crosswalk build scripts
This saves us from having to download a huge amount of files for
crosswalk. This is incredibly useful in places where there is a bandwidth
cap.

Note that this was obtained from a crosswalk build at branch "crosswalk-16"

The gitattributes file is there to make sure the .sh file line endings
aren't mangled. I could set them to binary, but I think setting shell
files to binary files is pretty terrible tbh.